+++ tests/assets/minimal_instance/training_config.yaml
