+++ tests/test_train.py
@@ -1,19 +1,20 @@
-from omegaconf import DictConfig
+from omegaconf import DictConfig, OmegaConf
 from pathlib import Path
 import sys
+import os
 
 import pytest
-from sleap_nn.train import main
+from sleap_nn.train import train, main
 
 
 @pytest.fixture
-def sample_cfg(sleap_data_dir, tmp_path):
+def sample_cfg(minimal_instance, tmp_path):
     config = DictConfig(
         {
             "data_config": {
                 "provider": "LabelsReader",
-                "train_labels_path": f"{sleap_data_dir}/minimal_instance.pkg.slp",
-                "val_labels_path": f"{sleap_data_dir}/minimal_instance.pkg.slp",
+                "train_labels_path": f"{minimal_instance}",
+                "val_labels_path": f"{minimal_instance}",
                 "user_instances_only": True,
                 "data_pipeline_fw": "torch_dataset",
                 "np_chunks_path": None,
@@ -107,7 +108,7 @@ def sample_cfg(sleap_data_dir, tmp_path):
                 "seed": 1000,
                 "use_wandb": False,
                 "save_ckpt": True,
-                "save_ckpt_path": f"{tmp_path}/test_cli_main",
+                "save_ckpt_path": (Path(tmp_path) / "test_cli_main").as_posix(),
                 "resume_ckpt_path": None,
                 "wandb": {
                     "entity": None,
@@ -135,6 +136,243 @@ def sample_cfg(sleap_data_dir, tmp_path):
         }
     )
     return config
+
+
+def test_train_method(minimal_instance, tmp_path: str):
+    train(
+        train_labels_path=minimal_instance,
+        val_labels_path=minimal_instance,
+        test_file_path=minimal_instance,
+        max_epochs=1,
+        trainer_accelerator="cpu",
+        head_configs="centered_instance",
+        save_ckpt=True,
+        save_ckpt_path=(Path(tmp_path) / "test_train_method").as_posix(),
+    )
+    folder_created = (Path(tmp_path) / "test_train_method").exists()
+    assert folder_created
+    assert (
+        (Path(tmp_path) / "test_train_method").joinpath("training_config.yaml").exists()
+    )
+    assert (Path(tmp_path) / "test_train_method").joinpath("best.ckpt").exists()
+    assert (Path(tmp_path) / "test_train_method").joinpath("pred_val.slp").exists()
+    assert (Path(tmp_path) / "test_train_method").joinpath("pred_test.slp").exists()
+
+    # with augmentations
+    with pytest.raises(ValueError):
+        train(
+            train_labels_path=minimal_instance,
+            val_labels_path=minimal_instance,
+            max_epochs=1,
+            trainer_accelerator="cpu",
+            head_configs="centered_instance",
+            use_augmentations_train=True,
+            intensity_aug="intensity",
+            geometry_aug=["rotation", "scale"],
+            save_ckpt=True,
+            save_ckpt_path=f"{tmp_path}/test_aug",
+        )
+
+    with pytest.raises(ValueError):
+        train(
+            train_labels_path=minimal_instance,
+            val_labels_path=minimal_instance,
+            max_epochs=1,
+            trainer_accelerator="cpu",
+            head_configs="centered_instance",
+            use_augmentations_train=True,
+            intensity_aug="uniform_noise",
+            geometry_aug="rotate",
+            save_ckpt=True,
+            save_ckpt_path=f"{tmp_path}/test_aug",
+        )
+
+    train(
+        train_labels_path=minimal_instance,
+        val_labels_path=minimal_instance,
+        max_epochs=1,
+        trainer_accelerator="cpu",
+        head_configs="centered_instance",
+        use_augmentations_train=True,
+        intensity_aug=["uniform_noise", "gaussian_noise", "contrast"],
+        geometry_aug=["rotation", "scale"],
+        save_ckpt=True,
+        save_ckpt_path=f"{tmp_path}/test_aug",
+    )
+
+    config = OmegaConf.load(f"{tmp_path}/test_aug/training_config.yaml")
+    assert config.data_config.augmentation_config.intensity.uniform_noise_p == 1.0
+    assert config.data_config.augmentation_config.intensity.gaussian_noise_p == 1.0
+    assert config.data_config.augmentation_config.intensity.contrast_p == 1.0
+    assert config.data_config.augmentation_config.intensity.brightness_p != 1.0
+    assert config.data_config.augmentation_config.geometric.affine_p == 1.0
+
+    train(
+        train_labels_path=minimal_instance,
+        val_labels_path=minimal_instance,
+        max_epochs=1,
+        trainer_accelerator="cpu",
+        head_configs="centered_instance",
+        use_augmentations_train=True,
+        intensity_aug="brightness",
+        geometry_aug=["translate", "erase_scale", "mixup"],
+        save_ckpt=True,
+        save_ckpt_path=f"{tmp_path}/test_aug",
+    )
+
+    config = OmegaConf.load(f"{tmp_path}/test_aug/training_config.yaml")
+    assert config.data_config.augmentation_config.intensity.uniform_noise_p == 0.0
+    assert config.data_config.augmentation_config.intensity.brightness_p == 1.0
+    assert config.data_config.augmentation_config.geometric.affine_p == 1.0
+    assert config.data_config.augmentation_config.geometric.erase_p == 1.0
+    assert config.data_config.augmentation_config.geometric.mixup_p == 1.0
+
+    ## test with passing dicts for aug
+    train(
+        train_labels_path=minimal_instance,
+        val_labels_path=minimal_instance,
+        max_epochs=1,
+        trainer_accelerator="cpu",
+        head_configs="centered_instance",
+        use_augmentations_train=True,
+        intensity_aug={
+            "uniform_noise_min": 0.0,
+            "uniform_noise_max": 1.0,
+            "uniform_noise_p": 1.0,
+        },
+        geometry_aug={"rotation": 180.0, "affine_p": 1.0},
+        save_ckpt=True,
+        save_ckpt_path=f"{tmp_path}/test_aug",
+    )
+
+    config = OmegaConf.load(f"{tmp_path}/test_aug/training_config.yaml")
+    assert config.data_config.augmentation_config.intensity.uniform_noise_p == 1.0
+    assert config.data_config.augmentation_config.geometric.affine_p == 1.0
+    assert config.data_config.augmentation_config.geometric.rotation == 180.0
+
+    # backbone configs #TODO
+    with pytest.raises(ValueError):
+        train(
+            train_labels_path=minimal_instance,
+            val_labels_path=minimal_instance,
+            max_epochs=1,
+            backbone_config="resnet",
+            trainer_accelerator="cpu",
+            head_configs="centroid",
+            save_ckpt=True,
+            save_ckpt_path=f"{tmp_path}/test_aug",
+        )
+
+    # head configs
+    with pytest.raises(ValueError):
+        train(
+            train_labels_path=minimal_instance,
+            val_labels_path=minimal_instance,
+            max_epochs=1,
+            backbone_config="unet",
+            trainer_accelerator="cpu",
+            head_configs="center",
+            save_ckpt=True,
+            save_ckpt_path=f"{tmp_path}/test_aug",
+        )
+
+    train(
+        train_labels_path=minimal_instance,
+        val_labels_path=minimal_instance,
+        max_epochs=20,
+        trainer_accelerator="cpu",
+        backbone_config="unet",
+        head_configs={
+            "single_instance": None,
+            "centered_instance": None,
+            "bottomup": None,
+            "centroid": {
+                "confmaps": {
+                    "anchor_part": None,
+                    "sigma": 1.5,
+                    "output_stride": 2,
+                }
+            },
+        },
+        save_ckpt=False,
+        save_ckpt_path=f"{tmp_path}/test_centroid",
+    )
+    config = OmegaConf.load(f"{tmp_path}/test_centroid/training_config.yaml")
+    assert config.model_config.head_configs.centroid is not None
+    assert config.model_config.head_configs.bottomup is None
+
+    # train(
+    #     train_labels_path=minimal_instance,
+    #     val_labels_path=minimal_instance,
+    #     max_epochs=1,
+    #     trainer_accelerator="cpu",
+    #     head_configs="single_instance",
+    #     save_ckpt=True,
+    #     save_ckpt_path=f"{tmp_path}/test_single_instabce",
+    #     lr_scheduler="ReduceLROnPlateau",
+    # )
+    # config = OmegaConf.load(f"{tmp_path}/test_single_instabce/training_config.yaml")
+    # assert config.model_config.head_configs.single_instance is not None
+    # assert config.model_config.head_configs.bottomup is None
+
+    train(
+        train_labels_path=minimal_instance,
+        val_labels_path=minimal_instance,
+        max_epochs=1,
+        trainer_accelerator="cpu",
+        head_configs="bottomup",
+        save_ckpt=True,
+        save_ckpt_path=f"{tmp_path}/test_bottomup",
+        lr_scheduler="StepLR",
+    )
+    config = OmegaConf.load(f"{tmp_path}/test_bottomup/training_config.yaml")
+    assert config.model_config.head_configs.bottomup is not None
+    assert config.model_config.head_configs.centroid is None
+
+    ## pass dict for head_configs
+    train(
+        train_labels_path=minimal_instance,
+        val_labels_path=minimal_instance,
+        max_epochs=1,
+        trainer_accelerator="cpu",
+        head_configs={
+            "centered_instance": {
+                "confmaps": {
+                    "part_names": None,
+                    "sigma": 2.5,
+                    "output_stride": 2,
+                    "anchor_part": None,
+                }
+            }
+        },
+        save_ckpt=True,
+        save_ckpt_path=f"{tmp_path}/test_custom_head",
+        lr_scheduler="StepLR",
+    )
+    config = OmegaConf.load(f"{tmp_path}/test_custom_head/training_config.yaml")
+    assert config.model_config.head_configs.centered_instance is not None
+    assert config.model_config.head_configs.centroid is None
+
+    ## pass dict for scheduler
+    train(
+        train_labels_path=minimal_instance,
+        val_labels_path=minimal_instance,
+        max_epochs=1,
+        trainer_accelerator="cpu",
+        head_configs={
+            "centroid": {
+                "confmaps": {"anchor_part": None, "sigma": 2.5, "output_stride": 2}
+            }
+        },
+        save_ckpt=False,
+        save_ckpt_path=f"{tmp_path}/test_scheduler",
+        lr_scheduler={
+            "scheduler": "StepLR",
+            "step_lr": {"step_size": 10, "gamma": 0.1},
+        },
+    )
+    config = OmegaConf.load(f"{tmp_path}/test_scheduler/training_config.yaml")
+    assert config.trainer_config.lr_scheduler.step_lr.step_size == 10
 
 
 @pytest.mark.skipif(
